Adjust: moved chart config into own file in utils
